requirements.txt
torchdiffeq==0.2.1
matplotlib==3.3.4
tqdm==4.56.2
scipy==1.6.0
seaborn==0.11.1
hickle==4.0.4
pandas==1.2.2
numpy==1.20.1
pillow==8.1.0
six==1.15.0
